Move auth-actions folder
