+++ app/(auth-pages)/reset-password/page.tsx
@@ -1,4 +1,4 @@
-import { resetPasswordAction } from '@/actions/auth-actions'
+import { resetPasswordAction } from '@/app/_actions/auth-actions'
 import Field from '@/components/form/field'
 import { FormMessage, Message } from '@/components/form/form-message'
 import { SubmitButton } from '@/components/submit-button'
