+++ components/navigation/header.tsx
@@ -1,7 +1,7 @@
 import Image from 'next/image'
 import Link from 'next/link'
 
-import { signOutAction } from '@/actions/auth-actions'
+import { signOutAction } from '@/app/_actions/auth-actions'
 import { getCurrentUser } from '@/actions/get-current-user'
 
 import { Button } from '../ui/button'
